moved existing experiment modules to package
